pyproject.toml
[project]
name = "auto-listing-bot"
version = "0.1.0"
description = "The Listing Bot scrapes products from various websites, stores them in Google Sheets, and lists them on connected WordPress stores."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bcrypt==4.0.1",
    "bs4>=0.0.2",
    "fastapi>=0.116.1",
    "gspread>=6.2.1",
    "httpx>=0.28.1",
    "openpyxl>=3.1.5",
    "pandas>=2.3.0",
    "passlib>=1.7.4",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.20",
    "requests>=2.32.4",
    "schedule>=1.2.2",
    "sqlalchemy>=2.0.42",
    "uvicorn>=0.35.0",
    "woocommerce>=3.0.0",
]

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
]
